+++ src/main/java/com/drmangotea/tfmg/content/electricity/base/ElectricalNetwork.java
@@ -102,7 +102,7 @@ public class ElectricalNetwork {
             if (member instanceof VoltageAlteringBlockEntity be) {
                 be.updateInFront();
             }
-            member.getLevelAccessor().blockUpdated(member.position(), member.getLevelAccessor().getBlockState(member.position()).getBlock());
+            member.getLevelAccessor().blockUpdated(member.getPos(), member.getLevelAccessor().getBlockState(member.getPos()).getBlock());
            // if (resistance != 0) {
            //     member.setNetworkResistance(1f / resistance);
            // } else member.setNetworkResistance(0);